feat(get): replace status command with get cluster subcommand
- add get cluster [NAME] for detailed health view with JSON support
- move formatState, checkmark, formatServices from status.go to get.go
- add json struct tags to Status, NodeStatus, NodeHealth, NetworkHealth,
  MountPoint, ServiceHealth
- remove standalone status command from root
- update completion tests and integration tests
