+++ cmd/sind/root.go
@@ -38,7 +38,6 @@ func NewRootCommand() *cobra.Command {
 	cmd.AddCommand(newCreateCommand())
 	cmd.AddCommand(newDeleteCommand())
 	cmd.AddCommand(newGetCommand())
-	cmd.AddCommand(newStatusCommand())
 	cmd.AddCommand(newPowerCommand())
 	cmd.AddCommand(newSSHCommand())
 	cmd.AddCommand(newEnterCommand())
